Kernel: Use `size_t` to keep track of the number of pages in a region
We were previously using a 32-bit unsigned integer for this, which
caused us to start truncating region sizes when multiplied with
`PAGE_SIZE` on hardware with a lot of memory.

+++ Kernel/Memory/PhysicalRegion.cpp
@@ -63,7 +63,7 @@ void PhysicalRegion::initialize_zones()
     make_zones(small_zone_size);
 }
 
-OwnPtr<PhysicalRegion> PhysicalRegion::try_take_pages_from_beginning(unsigned page_count)
+OwnPtr<PhysicalRegion> PhysicalRegion::try_take_pages_from_beginning(size_t page_count)
 {
     VERIFY(page_count > 0);
     VERIFY(page_count < m_pages);
